fixes some issues

+++ src/processor.cpp
@@ -3,14 +3,13 @@
 #include <thread>
 #include <chrono>
 
-
-float Processor::Utilization() {
+double Processor::Utilization() {
   auto prevIdle = LinuxParser::IdleJiffies();
   auto  prevTotal = LinuxParser::Jiffies();
   std::this_thread::sleep_for(std::chrono::milliseconds(200));
   auto idle = LinuxParser::IdleJiffies();
   auto total = LinuxParser::Jiffies();
-  auto diffTotal = static_cast<float>(total - prevTotal);
-  auto diffIdle =  static_cast<float>(idle - prevIdle);
+  auto diffTotal = static_cast<double >(total - prevTotal);
+  auto diffIdle =  static_cast<double>(idle - prevIdle);
   return (diffTotal - diffIdle)/diffTotal;
 }
